done upload image sb3

+++ index.php
@@ -16,7 +16,7 @@ $rows_per_page = 5;
 // Determine the starting row number for the current page
 $start= ($page-1)*$rows_per_page;
 
-$sql = "SELECT idea_tbl.IdeaTitle, category_tbl.CategoryTitle, user_tbl.Username, idea_tbl.DatePost, idea_tbl.IdeaDescription, idea_tbl.IdeaAnonymous from idea_tbl 
+$sql = "SELECT idea_tbl.IdeaId, idea_tbl.IdeaTitle, category_tbl.CategoryTitle, user_tbl.Username, idea_tbl.DatePost, idea_tbl.IdeaDescription, idea_tbl.IdeaAnonymous from idea_tbl 
 INNER JOIN user_tbl ON idea_tbl.UserId =user_tbl.UserId 
 INNER JOIN category_tbl ON idea_tbl.CategoryId= category_tbl.CategoryId 
 WHERE is_hidden=0 ORDER BY idea_tbl.IdeaId DESC LIMIT $start,$rows_per_page";
@@ -396,29 +396,57 @@ if(isset($_POST["submit_comment_post"])){
                 </div>
               </div>
 
-<?php
-      //displaying every ideas from database
-    while ($row = mysqli_fetch_assoc($result)){
-      echo'<div class="card">';
-      echo'<div class="card-body">';  
-        echo'<h1 class="card-title">'.$row['IdeaTitle'].'</h1>';
-        if($row['IdeaAnonymous'] == 0) {
-          echo '<h5 class="card-author">'.$row['Username'].'</h5>';
-        }
-        else if($row['IdeaAnonymous'] == 1){
-          echo '<h5 class="card-author">Anonymous</h5>';
-        }
+              <?php
+            //displaying every ideas from database
+            while ($row = mysqli_fetch_assoc($result)) {
+              echo '<div class="card">';
+              echo '<div class="card-body">';
+              echo '<h1 class="card-title">' . $row['IdeaTitle'] . '</h1>';
+              if ($row['IdeaAnonymous'] == 0) {
+                echo '<h5 class="card-author">' . $row['Username'] . '</h5>';
+              } else if ($row['IdeaAnonymous'] == 1) {
+                echo '<h5 class="card-author">Anonymous</h5>';
+              }
 
-        echo '<h5 class="card-category">'.$row['CategoryTitle'].'</h5>'; 
-              echo'<p class="card-text">'.$row['IdeaDescription'].'</p>';
-              
-             echo'<button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#viewUserIdea">See more</button>';
-             echo'<a href="#" class="btn btn-primary" style="background-color: darkcyan; margin-left: 5px;"><i class="bi bi-hand-thumbs-up"></i></a>';
-             echo'<a href="#" class="btn btn-primary" style="background-color: darkcyan; margin-left: 5px;"><i class="bi bi-hand-thumbs-down"></i></a>';
-            echo'</div>';
-          echo'</div>';
-}
- ?>                    
+              echo '<h5 class="card-category">' . $row['CategoryTitle'] . '</h5>';
+              echo '<p class="card-text">' . $row['IdeaDescription'] . '</p>';
+
+
+
+              $ideaid = $row['IdeaId'];
+              $imageidea_query = "SELECT IdeaImage FROM ideamedia_tbl WHERE IdeaId=$ideaid";
+              $imageidea_result = mysqli_query($dbconn, $imageidea_query);
+              $imageidea_count = mysqli_num_rows($imageidea_result);
+              if ($imageidea_count > 0) {
+                echo '<section class="pb-4">';
+                echo '    <div class="bg-white border rounded-5">';
+                echo '        <section class="p-4 d-flex w-100">';
+                echo '            <div class="lightbox" data-mdb-zoom-level="0.25" data-id="lightbox-8e0in48hs">';
+                echo '                <div class="row">';
+                while ($imageidea_row = mysqli_fetch_assoc($imageidea_result)) {
+                  $imageidea_path = '' . $imageidea_row['IdeaImage'];
+                  if (file_exists($imageidea_path)) {
+                    echo '   <div class="col-lg-4 mb-4">';
+                    echo '         <img src="' . $imageidea_path . '"  alt="idea image" class="w-100 shadow-1-strong rounded mb-4" style="width: 150px; height: 150px; object-fit: contain;">';
+                    echo '   </div>';
+                  }
+                }
+                echo '</div>';
+                echo '            </div>';
+                echo '        </section>';
+                echo '    </div>';
+                echo '</section>';
+              }
+
+
+
+              echo '<button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#viewUserIdea">See more</button>';
+              echo '<a href="#" class="btn btn-primary" style="background-color: darkcyan;"><i class="bi bi-hand-thumbs-up"></i></a>';
+              echo '<a href="#" class="btn btn-primary" style="background-color: darkcyan;"><i class="bi bi-hand-thumbs-down"></i></a>';
+              echo '</div>';
+              echo '</div>';
+            }
+            ?>                   
                 <!--
                   <form method="POST">
 						        <input type="submit" class="like_btn" name="like_btn" value="Like" />
